requirements.txt
python-telegram-bot==21.10
aiosqlite==0.20.0
